A lot of algo experimenting, but at the moment it's not working as well as before.
Some things are better. Adding a little bit of rotation to the velocity
vector killed off all oscillation.

Adjusting the lin-repulsion and local-repulsion to be a bit stronger on
closer elements helped.

Also if elements are too close there is no repulsion, so did a
corresponding adjustment to attraction. If they are too close, the
attraction actually pushes them apart instead. That seemed to be good.

I think a bit more gravity or gravity-time is needed to pull things together.