WIP on a forked builder mode.
This would function very similarly to a linear graph, but hopefully
would enable easier debugging.

The goal is that a linear graph can be used in production, but while
working on code that builds a graph, I can use this.

The difference is that at every point in time, the graph being built is
a forked graph. Mutation operations make it linear, do the mutation and
then back to forked. The new forked graph is stored in an atom.
Hopefully debug tools will be able to capture that atom to preserve
intermediate state.